Finish refactoring pytorch tests
